Move environment in use endpoint to environments namespace
